# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BizPortal Multi-Division API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /accounting/payments:
    get:
      operationId: listAccountingPayments
      tags: [accounting]
      summary: List manual payment records
      parameters:
        - in: query
          name: paymentType
          schema: { type: string, enum: [inbound, outbound] }
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
        - in: query
          name: sourceType
          schema: { type: string, nullable: true }
        - in: query
          name: sourceDocId
          schema: { type: integer, nullable: true }
        - in: query
          name: refDocNumber
          schema: { type: string, nullable: true }
        - in: query
          name: company
          schema: { type: string, nullable: true }
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AccountingPayment" }
    post:
      operationId: createAccountingPayment
      tags: [accounting]
      summary: Record a manual payment and auto-post the journal entry
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateAccountingPaymentBody" }
      responses:
        "201":
          description: Created payment with linked entry
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountingPaymentDetail" }
  /accounting/payments/{id}:
    get:
      operationId: getAccountingPayment
      tags: [accounting]
      summary: Get payment detail with linked journal entry
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Payment detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AccountingPaymentDetail" }
...
components:
  schemas:
...
    AccountingEntry:
      type: object
      properties:
        id: { type: integer }
        entryNumber: { type: string }
        journalId: { type: integer }
        date: { type: string }
        ref: { type: string, nullable: true }
        description: { type: string, nullable: true }
        status:
          type: string
          enum: [draft, posted]
        source:
          type: string
          enum: [manual, sales_invoice, purchase_bill, sales_payment, purchase_payment, pos_sale, ecommerce_order, stock_received, manual_payment]
        sourceId: { type: integer, nullable: true }
        totalDebit: { type: number }
        totalCredit: { type: number }
        createdById: { type: string, nullable: true }
        createdAt: { type: string }
      required: [id, entryNumber, journalId, date, status, source, totalDebit, totalCredit, createdAt]
    AccountingEntryLine:
      type: object
      properties:
        id: { type: integer }
        entryId: { type: integer }
        accountId: { type: integer }
        description: { type: string, nullable: true }
        debit: { type: number }
        credit: { type: number }
      required: [id, entryId, accountId, debit, credit]
...
    AccountingEntryDetail:
      allOf:
        - $ref: "#/components/schemas/AccountingEntry"
        - type: object
          properties:
            lines:
              type: array
              items: { $ref: "#/components/schemas/AccountingEntryLine" }
          required: [lines]
...
    AccountingPayment:
      type: object
      properties:
        id: { type: integer }
        paymentType:
          type: string
          enum: [inbound, outbound]
        status:
          type: string
          enum: [posted, voided]
        amount: { type: number }
        journalId: { type: integer }
        partnerName: { type: string, nullable: true }
        date: { type: string }
        ref: { type: string, nullable: true }
        memo: { type: string, nullable: true }
        entryId: { type: integer, nullable: true }
        voidEntryId: { type: integer, nullable: true }
        voidReason: { type: string, nullable: true }
        sourceType: { type: string, nullable: true }
        sourceDocId: { type: integer, nullable: true }
        createdById: { type: string, nullable: true }
        createdAt: { type: string }
      required: [id, paymentType, status, amount, journalId, date, createdAt]
    AccountingPaymentDetail:
      allOf:
        - $ref: "#/components/schemas/AccountingPayment"
        - type: object
          properties:
            entry:
              nullable: true
              allOf:
                - $ref: "#/components/schemas/AccountingEntryDetail"
    CreateAccountingPaymentBody:
      type: object
      properties:
        paymentType:
          type: string
          enum: [inbound, outbound]
        amount: { type: number }
        journalId: { type: integer }
        partnerName: { type: string }
        date: { type: string }
        ref: { type: string }
        memo: { type: string }
        sourceType: { type: string }
        sourceDocId: { type: integer }
      required: [paymentType, amount, journalId, date]
...